+++ web/packages/new/photos/components/DevSettings.tsx
@@ -1,3 +1,4 @@
+import { FocusVisibleButton } from "@/base/components/mui/FocusVisibleButton";
 import { useIsMobileWidth } from "@/base/hooks";
 import { ensureOk } from "@/base/http";
 import { getKVS, removeKV, setKV } from "@/base/kv";
@@ -18,7 +19,6 @@ import { useFormik } from "formik";
 import { t } from "i18next";
 import React, { useEffect, useState } from "react";
 import { z } from "zod";
-import { FocusVisibleButton } from "./mui/FocusVisibleButton";
 import { SlideUpTransition } from "./mui/SlideUpTransition";
 
 interface DevSettingsProps {
