+++ web/packages/new/photos/components/WhatsNew.tsx
@@ -1,3 +1,4 @@
+import { FocusVisibleButton } from "@/base/components/mui/FocusVisibleButton";
 import { ensureElectron } from "@/base/electron";
 import { useIsMobileWidth } from "@/base/hooks";
 import { ut } from "@/base/i18n";
@@ -14,7 +15,6 @@ import {
 } from "@mui/material";
 import React, { useEffect } from "react";
 import { didShowWhatsNew } from "../services/changelog";
-import { FocusVisibleButton } from "./mui/FocusVisibleButton";
 import { SlideUpTransition } from "./mui/SlideUpTransition";
 
 interface WhatsNewProps {
